refactor: urqlでgraphqlを叩く処理をapiディレクトリにまとめる
