+++ src/api/index.ts
@@ -1,0 +1,3 @@
+export * from "./usePokemonBaseStatsQuery";
+export * from "./usePokemonFormsQuery";
+export * from "./usePokemonIdQuery";
