--- src/features/forms/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./usePokemonFormsQuery";
