+++ src/features/search/components/container.tsx
@@ -2,12 +2,13 @@ import { useSetAtom } from "jotai";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { BehaviorSubject, debounceTime } from "rxjs";
 
+import { usePokemonIdQuery } from "@/api";
 import { pokemonIndividualIdAtom, pokemonSpeciesIdAtom } from "@/atoms";
 import { useErrorToast } from "@/hooks";
 
 import { Presentation } from "./presentation";
 import { Datalist } from "../datalist";
-import { useInputValue, usePokemonIdQuery } from "../hooks";
+import { useInputValue } from "../hooks";
 
 import type { FormEvent } from "react";
