+++ src/features/search/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from "./usePokemonIdQuery";
 export * from "./useInputValue";
