+++ src/features/status/components/index.tsx
@@ -2,20 +2,20 @@ import { HStack, Skeleton } from "@chakra-ui/react";
 import { useAtomValue } from "jotai";
 import { Suspense } from "react";
 
+import { usePokemonBaseStatsQuery } from "@/api";
 import { pokemonIndividualIdAtom } from "@/atoms";
 import { ADJUSTED, CURRENT } from "@/constants";
 import { useErrorToast } from "@/hooks";
 import { garchomp } from "@/mockData/pokemons";
 
 import { Status } from "./Status";
-import { usePokemonBaseStats } from "../hooks";
 
 export function StatusTableList() {
   const pokemonIndividualId = useAtomValue(pokemonIndividualIdAtom);
 
   const { showErrorToast } = useErrorToast();
 
-  const { baseStatsData, error } = usePokemonBaseStats(pokemonIndividualId);
+  const { baseStatsData, error } = usePokemonBaseStatsQuery(pokemonIndividualId);
 
   if (error) {
     showErrorToast({
