+++ src/features/status/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from "./usePokemonBaseStats";
 export * from "./useEffortValue";
